Moved readers to separate class. Deleted unfinished readers.
