+++ KeyReader/JsonStringKeyReader.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Nodes;
 
-namespace OpenAiConnector.KeyReaders;
+namespace KeyReader;
 
 public class JsonStringKeyReader : IApiKeyReader, IOrganizationIdReader
 {
