+++ Legba.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenAiConnector", "OpenAiCo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Legba.WPF", "Legba.WPF\Legba.WPF.csproj", "{601EDCF5-AEC7-4831-972D-58AE177694E8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KeyReader", "KeyReader\KeyReader.csproj", "{DD1A6FA1-7A3D-4FD7-815A-C6F9F4243342}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{601EDCF5-AEC7-4831-972D-58AE177694E8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{601EDCF5-AEC7-4831-972D-58AE177694E8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{601EDCF5-AEC7-4831-972D-58AE177694E8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DD1A6FA1-7A3D-4FD7-815A-C6F9F4243342}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DD1A6FA1-7A3D-4FD7-815A-C6F9F4243342}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DD1A6FA1-7A3D-4FD7-815A-C6F9F4243342}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DD1A6FA1-7A3D-4FD7-815A-C6F9F4243342}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
