--- OpenAiConnector/KeyReaders/AwsKmsKeyReader.cs
@@ -1,9 +1,0 @@
-﻿namespace OpenAiConnector.KeyReaders;
-
-public class AwsKmsKeyReader : IApiKeyReader, IOrganizationIdReader
-{
-    public string GetKey()
-    {
-        throw new NotImplementedException();
-    }
-}
