--- OpenAiConnector/KeyReaders/AzureKeyVaultKeyReader.cs
@@ -1,9 +1,0 @@
-﻿namespace OpenAiConnector.KeyReaders;
-
-public class AzureKeyVaultKeyReader : IApiKeyReader, IOrganizationIdReader
-{
-    public string GetKey()
-    {
-        throw new NotImplementedException();
-    }
-}
