--- OpenAiConnector/KeyReaders/GcpKmsKeyReader.cs
@@ -1,11 +1,0 @@
-﻿namespace OpenAiConnector.KeyReaders;
-
-public class GcpKmsKeyReader : IApiKeyReader, IOrganizationIdReader
-{
-    private readonly string _openAiApiKey;
-
-    public string GetKey()
-    {
-        return _openAiApiKey;
-    }
-}
